sells　controllerを削除
