+++ app/controllers/items_controller.rb
@@ -2,22 +2,30 @@ class ItemsController < ApplicationController
   def index
   end
 
-  # def create
-  #   # formのデータを受け取る
-  #   @item = Item.new(item_params)
-  # end
+  def new
+    @item = Item.new
+    @parents = Category.all.order("id ASC").limit(10)
+  end
+  def create
+    # formのデータを受け取る
+    @item = Item.new(item_params)
+    if @item.save
+      render 'index'
+    else
+      render 'new'
+    end
+  end
 
-  # private
+  private
 
-  # def item_params
-  #   # ストロングパラメータ
-  #   params.require(:item).
-  #     permit(:name, :image, :category, :brand, :size, :condition, :cost,
-  #     :shipping, :area, :date, :price, :buyer_id, :saler_id)
-  # end
+  def item_params
+    # ストロングパラメータ
+    params.require(:item).permit(:name, :image, :category, :brand, :size, :condition,
+       :cost, :shipping, :area, :date, :price, :buyer_id, :saler_id)
+  end
 
-  # def set_item
-  #   #データの取得
-  #   @item = Item.find(params[:id])
-  # end
+  def set_item
+    #データの取得
+    @item = Item.find(params[:id])
+  end
 end
