--- app/controllers/sells_controller.rb
@@ -1,31 +1,0 @@
-class SellsController < ApplicationController
-  def index
-  end
-
-  def new
-    @item = Item.new
-    @parents = Category.all.order("id ASC").limit(10)
-  end
-  def create
-    # formのデータを受け取る
-    @item = Item.new(item_params)
-    if @item.save
-      render 'index'
-    else
-      render 'new'
-    end
-  end
-
-  private
-
-  def item_params
-    # ストロングパラメータ
-    params.require(:item).permit(:name, :image, :category, :brand, :size, :condition,
-       :cost, :shipping, :area, :date, :price, :buyer_id, :saler_id)
-  end
-
-  def set_item
-    #データの取得
-    @item = Item.find(params[:id])
-  end
-end
