--- app/helpers/sells_helper.rb
@@ -1,2 +1,0 @@
-module SellsHelper
-end
