--- test/controllers/sells_controller_test.rb
@@ -1,7 +1,0 @@
-require 'test_helper'
-
-class SellsControllerTest < ActionDispatch::IntegrationTest
-  # test "the truth" do
-  #   assert true
-  # end
-end
